Changed partitioning field in length-partitioned inverted indexes from integer to short.
git-svn-id: https://hyracks.googlecode.com/svn/branches/hyracks_lsm_length_filter@2486 123451ca-8445-de46-9d55-352943316053

+++ hyracks-storage-am-lsm-invertedindex/src/main/java/edu/uci/ics/hyracks/storage/am/lsm/invertedindex/api/IInvertedIndexSearchModifier.java
@@ -15,13 +15,12 @@
 
 package edu.uci.ics.hyracks.storage.am.lsm.invertedindex.api;
 
-
 public interface IInvertedIndexSearchModifier {
     public int getOccurrenceThreshold(int numQueryTokens);
 
     public int getNumPrefixLists(int occurrenceThreshold, int numInvLists);
-    
-    public int getNumTokensLowerBound(int numQueryTokens);
-    
-    public int getNumTokensUpperBound(int numQueryTokens);
+
+    public short getNumTokensLowerBound(short numQueryTokens);
+
+    public short getNumTokensUpperBound(short numQueryTokens);
 }
